+++ src/react/components/editabletable/XUIEditableTable.d.ts
@@ -4,6 +4,10 @@ import { XUIEditableTableContextShape } from './contexts/XUIEditableTableContext
 
 interface BaseProps {
   /**
+   * Internal use only, used to recognize it's an editable or read-only table.
+   */
+  _variant?: 'editable' | 'readonly';
+  /**
    * A non-visible description of the table for accessibility purposes. Particularly useful
    * for scrollable tables, to help screenreaders understand the scrollable element.
    */
@@ -110,7 +114,12 @@ interface BaseProps {
   onReorderRow?: (startIndex: number, destinationIndex: number) => void;
   qaHook?: string;
   rowOptions?: XUIEditableTableContextShape['rowOptions'];
+  scrollContainerRef?: React.Ref<HTMLDivElement>;
   style?: React.CSSProperties;
+  /**
+   * Custom class(es) to add to the table element.
+   */
+  tableClassName?: string;
   /**
    * Validation message to show under the table if `isInvalid` is true.
    */