+++ src/react/components/editabletable/XUIEditableTable.js
@@ -1,116 +1,138 @@
-import React, { useLayoutEffect, useState } from 'react';
+import React, { useState } from 'react';
 import PropTypes from 'prop-types';
 import { nanoid } from 'nanoid';
 import invalid from '@xero/xui-icon/icons/invalid';
 import cn from 'classnames';
 
-import { tableName } from './private/constants';
+import { tableVariantClassNames } from './private/constants';
 import EditableTableColGroup from './private/EditableTableColGroup';
 import EditableTableWrapper from './private/EditableTableWrapper';
 import conditionallyRequiredValidator from '../helpers/conditionallyRequiredValidator';
 import XUIIcon from '../icon/XUIIcon';
 import { generateIdsFromControlId } from '../controlwrapper/helpers';
+import combineRefs from '../helpers/combineRefs';
 
-const XUIEditableTable = ({
-  ariaLabel,
-  children,
-  className,
-  columnWidths = [],
-  dndDragCancelledMessage,
-  dndDragOutsideMessage,
-  dndDragStartMessage,
-  dndDragUpdateMessage,
-  dndDropFailedMessage,
-  dndDropMessage,
-  dndInstructions = '',
-  hasPinnedFirstColumn,
-  hasPinnedLastColumn,
-  hiddenColumns,
-  id,
-  isInvalid,
-  maxWidth,
-  minWidth,
-  onReorderRow,
-  qaHook,
-  rowOptions,
-  style,
-  validationMessage,
-  ...spreadProps
-}) => {
-  const tableRef = React.useRef();
-  // Ensures the table id is only generated once, but changes if the prop changes.
-  const [calculatedId, setId] = useState(id || `${tableName}-${nanoid(10)}`);
+const XUIEditableTable = React.forwardRef(
+  (
+    {
+      _variant = 'editable',
+      ariaLabel,
+      children,
+      className,
+      columnWidths = [],
+      dndDragCancelledMessage,
+      dndDragOutsideMessage,
+      dndDragStartMessage,
+      dndDragUpdateMessage,
+      dndDropFailedMessage,
+      dndDropMessage,
+      dndInstructions = '',
+      hasPinnedFirstColumn,
+      hasPinnedLastColumn,
+      hiddenColumns,
+      id,
+      isInvalid,
+      maxWidth,
+      minWidth,
+      onReorderRow,
+      qaHook,
+      rowOptions,
+      scrollContainerRef,
+      style,
+      tableClassName,
+      validationMessage,
+      ...spreadProps
+    },
+    ref,
+  ) => {
+    const tableRef = React.useRef();
+    const combinedRef = combineRefs(ref, tableRef);
 
-  const wrapperIds = generateIdsFromControlId(calculatedId);
+    const tableVariantClassName = tableVariantClassNames[_variant];
+    // Ensures the table id is only generated once, but changes if the prop changes.
+    const [calculatedId, setId] = useState(id || `${tableVariantClassName}-${nanoid(10)}`);
 
-  return (
-    <>
-      {hiddenColumns && hiddenColumns.length > 0 && (
-        <style>
-          {hiddenColumns.map(
-            hiddenColumn =>
-              `#${calculatedId} .xui-editabletablerow > *:nth-child(${
-                parseInt(hiddenColumn) + 1
-              }) { display: none; }`,
-          )}
-        </style>
-      )}
-      <EditableTableWrapper
-        className={className}
-        columnWidths={columnWidths}
-        dndDragCancelledMessage={dndDragCancelledMessage}
-        dndDragOutsideMessage={dndDragOutsideMessage}
-        dndDragStartMessage={dndDragStartMessage}
-        dndDragUpdateMessage={dndDragUpdateMessage}
-        dndDropFailedMessage={dndDropFailedMessage}
-        dndDropMessage={dndDropMessage}
-        dndInstructions={dndInstructions}
-        hasPinnedFirstColumn={hasPinnedFirstColumn}
-        hasPinnedLastColumn={hasPinnedLastColumn}
-        isInvalid={isInvalid}
-        maxWidth={maxWidth}
-        minWidth={minWidth}
-        onReorderRow={onReorderRow}
-        rowOptions={rowOptions}
-        tableRef={tableRef}
-        validationMessage={validationMessage}
-      >
-        <table
-          {...spreadProps}
-          aria-describedby={isInvalid && validationMessage ? wrapperIds.message : undefined}
-          aria-invalid={isInvalid}
-          aria-label={ariaLabel}
-          className={cn(
-            tableName,
-            hasPinnedFirstColumn && `${tableName}-pinfirst`,
-            hasPinnedLastColumn && `${tableName}-pinlast`,
-          )}
-          data-automationid={qaHook}
-          id={wrapperIds.control}
-          ref={tableRef}
+    const wrapperIds = generateIdsFromControlId(calculatedId);
+
+    return (
+      <>
+        {hiddenColumns && hiddenColumns.length > 0 && (
+          <style>
+            {hiddenColumns.map(
+              hiddenColumn =>
+                `#${calculatedId} .xui-editabletablerow > *:nth-child(${
+                  parseInt(hiddenColumn) + 1
+                }) { display: none; }`,
+            )}
+          </style>
+        )}
+        <EditableTableWrapper
+          className={className}
+          columnWidths={columnWidths}
+          dndDragCancelledMessage={dndDragCancelledMessage}
+          dndDragOutsideMessage={dndDragOutsideMessage}
+          dndDragStartMessage={dndDragStartMessage}
+          dndDragUpdateMessage={dndDragUpdateMessage}
+          dndDropFailedMessage={dndDropFailedMessage}
+          dndDropMessage={dndDropMessage}
+          dndInstructions={dndInstructions}
+          hasPinnedFirstColumn={hasPinnedFirstColumn}
+          hasPinnedLastColumn={hasPinnedLastColumn}
+          isInvalid={isInvalid}
+          maxWidth={maxWidth}
+          minWidth={minWidth}
+          onReorderRow={onReorderRow}
+          rowOptions={rowOptions}
+          scrollContainerRef={scrollContainerRef}
+          tableClassName={tableVariantClassName}
+          tableRef={tableRef}
+          validationMessage={validationMessage}
         >
-          <EditableTableColGroup columnWidths={columnWidths} />
-          {children}
-        </table>
-      </EditableTableWrapper>
-      {isInvalid && validationMessage && (
-        <div className={`${tableName}--validation`} id={wrapperIds.message}>
-          <XUIIcon icon={invalid} />
-          {validationMessage}
-        </div>
-      )}
-    </>
-  );
-};
+          <table
+            {...spreadProps}
+            aria-describedby={isInvalid && validationMessage ? wrapperIds.message : undefined}
+            aria-invalid={isInvalid}
+            aria-label={ariaLabel}
+            className={cn(
+              tableVariantClassName,
+              hasPinnedFirstColumn && `${tableVariantClassName}-pinfirst`,
+              hasPinnedLastColumn && `${tableVariantClassName}-pinlast`,
+              tableClassName,
+            )}
+            data-automationid={qaHook}
+            id={wrapperIds.control}
+            ref={combinedRef}
+            style={style}
+          >
+            <EditableTableColGroup columnWidths={columnWidths} />
+            {children}
+          </table>
+        </EditableTableWrapper>
+        {isInvalid && validationMessage && (
+          <div className={`${tableVariantClassName}--validation`} id={wrapperIds.message}>
+            <XUIIcon icon={invalid} />
+            {validationMessage}
+          </div>
+        )}
+      </>
+    );
+  },
+);
 
 XUIEditableTable.propTypes = {
+  /**
+   * @ignore
+   * Internal use only, used to recognize it's an editable or read-only table
+   */
+  _variant: PropTypes.oneOf(['editable', 'readonly']),
+
   /**
    * A non-visible description of the table for accessibility purposes. Particularly useful
    * for scrollable tables, to help screenreaders understand the scrollable element.
    */
   ariaLabel: PropTypes.string,
 
-  children: PropTypes.oneOfType([PropTypes.element, PropTypes.arrayOf(PropTypes.element)]),
+  children: PropTypes.node,
   className: PropTypes.string,
 
   /**
@@ -326,7 +348,17 @@ XUIEditableTable.propTypes = {
     },
   }),
 
+  scrollContainerRef: PropTypes.oneOfType([
+    PropTypes.func,
+    PropTypes.shape({ current: PropTypes.instanceOf(Element) }),
+  ]),
+
   style: PropTypes.object,
+
+  /**
+   * Custom class(es) to add to the table element.
+   */
+  tableClassName: PropTypes.string,
 
   /**
    * Validation message to show under the table if `isInvalid` is true.