+++ src/react/components/editabletable/XUIEditableTableBody.js
@@ -3,14 +3,14 @@ import PropTypes from 'prop-types';
 import cn from 'classnames';
 
 import XUIEditableTableContext from './contexts/XUIEditableTableContext';
-import { tableName } from './private/constants';
-
-const baseName = `${tableName}body`;
+import XUIEditableTableClassContext from './contexts/XUIEditableTableClassContext';
 
 const XUIEditableTableBody = ({ children, className, qaHook, ...spreadProps }) => {
   const {
     dragAndDrop: { rowPlaceholder },
   } = React.useContext(XUIEditableTableContext);
+  const tableClassName = React.useContext(XUIEditableTableClassContext);
+  const baseName = `${tableClassName}body`;
 
   return (
     <tbody className={cn(baseName, className)} data-automationid={qaHook} {...spreadProps}>
@@ -21,7 +21,7 @@ const XUIEditableTableBody = ({ children, className, qaHook, ...spreadProps }) =
 };
 
 XUIEditableTableBody.propTypes = {
-  children: PropTypes.oneOfType([PropTypes.element, PropTypes.arrayOf(PropTypes.element)]),
+  children: PropTypes.node,
   className: PropTypes.string,
   qaHook: PropTypes.string,
 };