+++ src/react/components/editabletable/XUIEditableTableCellIconButton.js
@@ -2,11 +2,11 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import cn from 'classnames';
 
-import { tableName } from './private/constants';
+import { tableVariantClassNames } from './private/constants';
 import XUIEditableTableCell from './XUIEditableTableCell';
 import XUIIconButton from '../button/XUIIconButton';
 
-const baseName = `${tableName}celliconbutton`;
+const baseName = `${tableVariantClassNames.editable}celliconbutton`;
 
 const XUIEditableTableCellIconButton = ({
   ariaLabel,