+++ src/react/components/editabletable/XUIEditableTableFoot.js
@@ -2,9 +2,9 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import cn from 'classnames';
 
-import { tableName } from './private/constants';
+import { tableVariantClassNames } from './private/constants';
 
-const baseName = `${tableName}foot`;
+const baseName = `${tableVariantClassNames.editable}foot`;
 
 const XUIEditableTableFoot = ({ children, className, qaHook, ...spreadProps }) => {
   return (
@@ -15,7 +15,7 @@ const XUIEditableTableFoot = ({ children, className, qaHook, ...spreadProps }) =
 };
 
 XUIEditableTableFoot.propTypes = {
-  children: PropTypes.oneOfType([PropTypes.element, PropTypes.arrayOf(PropTypes.element)]),
+  children: PropTypes.node,
   className: PropTypes.string,
   qaHook: PropTypes.string,
 };