+++ src/react/components/editabletable/private/constants.ts
@@ -1,9 +1,11 @@
 import { ns } from '../../helpers/xuiClassNamespace';
 
-export const tableName = `${ns}-editabletable`;
-export const wrapperName = `${tableName}wrapper`;
-
 export const borderSpacing = 1;
 
 // TODO: Replace references to this with $xui-control-size-standard when SASS tokens can be shared with React
 export const xuiControlSizeStandard = '40px';
+
+export const tableVariantClassNames = {
+  readonly: `${ns}-readonlytable`,
+  editable: `${ns}-editabletable`,
+};