+++ src/react/components/table/__tests__/table-test.js
@@ -1,6 +1,7 @@
 import React from 'react';
-import Enzyme, { mount } from 'enzyme';
+import Enzyme, { mount, render } from 'enzyme';
 import Adapter from '@wojtekmaj/enzyme-adapter-react-17';
+import toJson from 'enzyme-to-json';
 import { axe, toHaveNoViolations } from 'jest-axe';
 import renderer from 'react-test-renderer';
 import { nanoid } from 'nanoid';
@@ -10,6 +11,11 @@ import { variations } from '../stories/variations';
 jest.mock('nanoid');
 nanoid.mockImplementation(() => 'testDropdownId');
 
+jest.mock('react', () => ({
+  ...jest.requireActual('react'),
+  useLayoutEffect: jest.requireActual('react').useEffect,
+}));
+
 Enzyme.configure({ adapter: new Adapter() });
 expect.extend(toHaveNoViolations);
 
@@ -17,10 +23,11 @@ describe('<XUITable />', () => {
   describe('emulate stories', () => {
     variations.forEach(({ storyKind, storyTitle, examples }) => {
       it(`should render scenario "${storyKind} ${storyTitle}" correctly`, () => {
-        const Comparison = examples.map(TestScaffold);
-        const component = renderer.create(<div>{Comparison}</div>);
+        // Arrange
+        const wrapper = render(<div>{examples.map(TestScaffold)}</div>);
 
-        expect(component).toMatchSnapshot();
+        // Assert
+        expect(toJson(wrapper)).toMatchSnapshot();
       });
     });
   });